Update README.md
This project is a comprehensive analysis of HR datasets to extract valuable insights to support data-driven decision-making in human resources management. As part of my graduation project for DEPI, I performed an end-to-end analysis, including data cleaning, exploratory data analysis (EDA), statistical analysis, and visualization. The goal was to uncover trends, patterns, and key metrics related to employee performance, attrition, recruitment, diversity, and other HR-related factors.
The project includes the development of an interactive dashboard using tools like Power BI and Tableau, which provides stakeholders with an intuitive and dynamic way to explore the data.